requirements.txt
# --- Core AI / Face Recognition ---
onnxruntime-gpu==1.16.3
insightface==0.7.3
opencv-python==4.8.1.78
numpy==1.26.2
scipy==1.11.4

# --- Vector Search ---
faiss-cpu==1.7.4

# --- GUI ---
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0

# --- Database (SQL Server) ---
pyodbc==5.0.1
sqlalchemy==2.0.23

# --- Reports & Data ---
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.8

# --- API Server & Networking (Edge Computing) ---
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
requests>=2.31.0

# --- Utilities ---
python-dotenv==1.0.0
Pillow==10.1.0
psutil==5.9.7
loguru==0.7.2

# --- Dev / Testing ---
pytest==7.4.3
pytest-qt==4.2.0

# ============================================================
# GHI CHÚ CÀI ĐẶT CHO GPU:
# 1. Cài CUDA Toolkit 11.8
# 2. Cài cuDNN 8.9.x (cho CUDA 11.x)
# 3. Thêm đường dẫn CUDA vào Environment Variables (Path)
# 4. Chạy: pip install onnxruntime-gpu==1.16.3
#
# GHI CHÚ MÔI TRƯỜNG BIÊN (EDGE CLIENT & ANTI-SPOOFING):
# - Để chạy chức năng chống gian lận trên CPU mà không bị lỗi biên dịch:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
# ============================================================